# cmd/patient-api/docs/swagger.yaml
basePath: /patient/api
consumes:
- application/json
definitions:
  datastore.BloodPressure:
    properties:
      created_at:
        type: string
      date_time:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      diastolic:
        type: integer
      id:
        type: integer
      patient_id:
        type: integer
      pulse:
        type: integer
      systolic:
        type: integer
      updated_at:
        type: string
    type: object
  datastore.Glucose:
    properties:
      created_at:
        type: string
      date_time:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      is_before_meal:
        type: boolean
      patient_id:
        type: integer
      updated_at:
        type: string
      value:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handler.AddCreditCardRequest:
    properties:
      card_token:
        type: string
    required:
    - card_token
    type: object
  handler.BloodPressureRequest:
    properties:
      date_time:
        type: string
      diastolic:
        type: integer
      pulse:
        type: integer
      systolic:
        type: integer
    required:
    - date_time
    - diastolic
    - pulse
    - systolic
    type: object
  handler.GlucoseRequest:
    properties:
      date_time:
        type: string
      is_before_meal:
        type: boolean
      value:
        type: integer
    required:
    - date_time
    - is_before_meal
    - value
    type: object
  handler.SigninRequest:
    properties:
      credential:
        type: string
    required:
    - credential
    type: object
  handler.SigninResponse:
    properties:
      phone_number:
        type: string
    type: object
  handler.VerifyOTPRequest:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
  handler.VerifyOTPResponse:
    properties:
      token:
        type: string
    type: object
  payment.Card:
    properties:
      brand:
        type: string
      default:
        type: boolean
      id:
        type: string
      last_digits:
        type: string
    type: object
  server.ErrorResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is a Synthia patient backend API.
  title: Synthia Patient Backend API
  version: 1.0.0
paths:
  /auth/signin:
    post:
      description: Initiate auth process with government credential which will sent
        OTP to patient's phone number
      parameters:
      - description: Patient government credential (Passport ID or National ID)
        in: body
        name: SigninRequest
        required: true
        schema:
          $ref: '#/definitions/handler.SigninRequest'
      responses:
        "201":
          description: OTP is sent to patient's phone number
          schema:
            $ref: '#/definitions/handler.SigninResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Provided credential is not in the hospital system
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Start signing-in with government credential
      tags:
      - Auth
  /auth/verify:
    post:
      description: Complete auth process with OTP verification. It will return token
        if verification success.
      parameters:
      - description: OTP that is sent to patient's phone number
        in: body
        name: VerifyOTPRequest
        required: true
        schema:
          $ref: '#/definitions/handler.VerifyOTPRequest'
      responses:
        "201":
          description: JWS Token for later use
          schema:
            $ref: '#/definitions/handler.VerifyOTPResponse'
        "400":
          description: OTP is invalid or expired
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Verify OTP and get token
      tags:
      - Auth
  /measurement/blood-pressure:
    post:
      parameters:
      - description: Blood pressure information
        in: body
        name: BloodPressureRequest
        required: true
        schema:
          $ref: '#/definitions/handler.BloodPressureRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/datastore.BloodPressure'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - UserID: []
      - JWSToken: []
      summary: Record blood pressure
      tags:
      - Measurement
  /measurement/glucose:
    post:
      parameters:
      - description: Glucose level information
        in: body
        name: GlucoseRequest
        required: true
        schema:
          $ref: '#/definitions/handler.GlucoseRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/datastore.Glucose'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - UserID: []
      - JWSToken: []
      summary: Record glucose level
      tags:
      - Measurement
  /payment/credit-card:
    get:
      responses:
        "200":
          description: List of saved cards
          schema:
            items:
              $ref: '#/definitions/payment.Card'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - UserID: []
      - JWSToken: []
      summary: Get lists of saved credit cards
      tags:
      - Payment
    post:
      parameters:
      - description: Token from Omise
        in: body
        name: AddCreditCardRequest
        required: true
        schema:
          $ref: '#/definitions/handler.AddCreditCardRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Failed to add credit card
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      security:
      - UserID: []
      - JWSToken: []
      summary: Add new credit card
      tags:
      - Payment
produces:
- application/json
securityDefinitions:
  JWSToken:
    in: header
    name: Authorization
    type: apiKey
  UserID:
    in: header
    name: X-USER-ID
    type: apiKey
swagger: "2.0"